Large simplification * No NumFieldsMonitor * No named URP; anonymous class is fine * URP need not be inserted in the happy path!
